audio: add moving detection and change signal data type to double

+++ modules/audio/common/audio_info.cc
@@ -42,13 +42,13 @@ void AudioInfo::InsertChannelData(const std::size_t index,
                                   const ChannelData& channel_data,
                                   const MicrophoneConfig& microphone_config) {
   while (index >= signals_.size()) {
-    signals_.push_back(std::deque<float>());
+    signals_.push_back(std::deque<double>());
   }
   int width = microphone_config.sample_width();
   const std::string& data = channel_data.data();
   for (int i = 0; i < channel_data.size(); i += width) {
     int16_t signal = ((int16_t(data[i])) << 8) | (0x00ff & data[i + 1]);
-    signals_[index].push_back(static_cast<float>(signal));
+    signals_[index].push_back(static_cast<double>(signal));
   }
   std::size_t max_signal_length = static_cast<std::size_t>(
       FLAGS_cache_signal_time * microphone_config.sample_rate());
@@ -57,14 +57,15 @@ void AudioInfo::InsertChannelData(const std::size_t index,
   }
 }
 
-std::vector<std::vector<float>> AudioInfo::GetSignals(const int signal_length) {
-  std::vector<std::vector<float>> signals;
+std::vector<std::vector<double>> AudioInfo::GetSignals(
+    const int signal_length) {
+  std::vector<std::vector<double>> signals;
   for (std::size_t i = 0; i < signals_.size(); ++i) {
     int start_index = static_cast<int>(signals_[i].size()) - signal_length;
     start_index = std::max(0, start_index);
-    std::deque<float>::iterator iter = signals_[i].begin();
+    std::deque<double>::iterator iter = signals_[i].begin();
     iter += start_index;
-    std::vector<float> signal(iter, signals_[i].end());
+    std::vector<double> signal(iter, signals_[i].end());
     signals.push_back(signal);
   }
   return signals;
